// apps/SnD/src/types/ElementTypes.ts
/* eslint-disable no-shadow */
enum ElementTypes {
    CHECK_BOX = 'checkbox',
    SELECT = 'select',
    MULTI_SELECT = 'multiSelect',
    LIST = 'list',
    GAP = 'gap',
    CSV = 'csv',
    AOTU_COMPLETE_SERVER_SIDE = 'asyncAutoComplete',
    AUTO_COMPLETE_CLIENT_SIDE = 'autoComplete',
    TEXT_AREA = 'textarea',
    TEXT = 'text',
    SELECT_SEARCH = 'select-search',
    MULTI_SELECT_SEARCH = 'multiselect-search',
    ICON_PICKER = 'icon-picker',
    DATE_PICKER = 'date-picker',
    DROP_ZONE = 'drop-zone',
    PHONE = 'phone'
}

export default ElementTypes;